Add `DDLQueryBuilder::class`, `DMLQueryBuilder::class`.
