--- src/Command/DMLCommand.php
@@ -1,150 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Db\Command;
-
-use InvalidArgumentException;
-use JsonException;
-use Yiisoft\Db\Exception\Exception;
-use Yiisoft\Db\Exception\InvalidConfigException;
-use Yiisoft\Db\Exception\NotSupportedException;
-use Yiisoft\Db\Expression\ExpressionInterface;
-use Yiisoft\Db\Query\Query;
-use Yiisoft\Db\Query\QueryBuilderInterface;
-use Yiisoft\Db\Schema\QuoterInterface;
-
-abstract class DMLCommand
-{
-    public function __construct(private QueryBuilderInterface $queryBuilder, private QuoterInterface $quoter)
-    {
-    }
-
-    /**
-     * Creates an INSERT SQL statement.
-     *
-     * For example,
-     *
-     * ```php
-     * $sql = $queryBuilder->insert('user', [
-     *     'name' => 'Sam',
-     *     'age' => 30,
-     * ], $params);
-     * ```
-     *
-     * The method will properly escape the table and column names.
-     *
-     * @param string $table the table that new rows will be inserted into.
-     * @param array|Query $columns the column data (name => value) to be inserted into the table or instance of
-     * {@see Query} to perform INSERT INTO ... SELECT SQL statement. Passing of {@see Query}.
-     * @param array $params the binding parameters that will be generated by this method. They should be bound to the
-     * DB command later.
-     *
-     * @throws Exception|InvalidArgumentException|InvalidConfigException|NotSupportedException
-     *
-     * @return string the INSERT SQL.
-     */
-    public function insert(string $table, Query|array $columns, array &$params = []): string
-    {
-        [$names, $placeholders, $values, $params] = $this->queryBuilder->prepareInsertValues($table, $columns, $params);
-
-        return 'INSERT INTO '
-            . $this->quoter->quoteTableName($table)
-            . (!empty($names) ? ' (' . implode(', ', $names) . ')' : '')
-            . (!empty($placeholders) ? ' VALUES (' . implode(', ', $placeholders) . ')' : $values);
-    }
-
-    /**
-     * Creates a SQL statement for resetting the sequence value of a table's primary key.
-     *
-     * The sequence will be reset such that the primary key of the next new row inserted will have the specified value
-     * or 1.
-     *
-     * @param string $tableName the name of the table whose primary key sequence will be reset.
-     * @param array|int|string|null $value the value for the primary key of the next new row inserted. If this is not
-     * set, the next new row's primary key will have a value 1.
-     *
-     * @throws Exception|NotSupportedException if this is not supported by the underlying DBMS.
-     *
-     * @return string the SQL statement for resetting sequence.
-     */
-    public function resetSequence(string $tableName, array|int|string|null $value = null): string
-    {
-        throw new NotSupportedException(static::class . ' does not support resetting sequence.');
-    }
-
-    /**
-     * Creates an UPDATE SQL statement.
-     *
-     * For example,
-     *
-     * ```php
-     * $params = [];
-     * $sql = $queryBuilder->update('user', ['status' => 1], 'age > 30', $params);
-     * ```
-     *
-     * The method will properly escape the table and column names.
-     *
-     * @param string $table the table to be updated.
-     * @param array $columns the column data (name => value) to be updated.
-     * @param array|string $condition the condition that will be put in the WHERE part. Please refer to
-     * {@see Query::where()} on how to specify condition.
-     * @param array $params the binding parameters that will be modified by this method so that they can be bound to the
-     * DB command later.
-     *
-     * @psalm-param array<string, ExpressionInterface|string> $columns
-     *
-     * @throws Exception|InvalidArgumentException
-     *
-     * @return string the UPDATE SQL.
-     */
-    public function update(string $table, array $columns, array|string $condition, array &$params = []): string
-    {
-        /**
-         * @var array $lines
-         * @var array $params
-         */
-        [$lines, $params] = $this->queryBuilder->prepareUpdateSets($table, $columns, $params);
-        $sql = 'UPDATE ' . $this->quoter->quoteTableName($table) . ' SET ' . implode(', ', $lines);
-        $where = $this->queryBuilder->buildWhere($condition, $params);
-
-        return $where === '' ? $sql : $sql . ' ' . $where;
-    }
-
-    /**
-     * Creates an SQL statement to insert rows into a database table if they do not already exist (matching unique
-     * constraints), or update them if they do.
-     *
-     * For example,
-     *
-     * ```php
-     * $sql = $queryBuilder->upsert('pages', [
-     *     'name' => 'Front page',
-     *     'url' => 'http://example.com/', // url is unique
-     *     'visits' => 0,
-     * ], [
-     *     'visits' => new Expression('visits + 1'),
-     * ], $params);
-     * ```
-     *
-     * The method will properly escape the table and column names.
-     *
-     * @param string $table the table that new rows will be inserted into/updated in.
-     * @param array|Query $insertColumns the column data (name => value) to be inserted into the table or instance of
-     * {@see Query} to perform `INSERT INTO ... SELECT` SQL statement.
-     * @param bool|array $updateColumns the column data (name => value) to be updated if they already exist. If `true`
-     * is passed, the column data will be updated to match the insert column data. If `false` is passed, no update will
-     * be performed if the column data already exists.
-     * @param array $params the binding parameters that will be generated by this method. They should be bound to the DB
-     * command later.
-     *
-     * @throws Exception|InvalidConfigException|JsonException|NotSupportedException if this is not supported by the
-     * underlying DBMS.
-     *
-     * @return string the resulting SQL.
-     */
-    public function upsert(string $table, Query|array $insertColumns, bool|array $updateColumns, array &$params): string
-    {
-        throw new NotSupportedException(static::class . ' does not support upsert.');
-    }
-}
